+++ src/history_to_hash.rs
@@ -20,30 +20,9 @@ pub fn read_history() -> HashMap<PathBuf, DirInfo> {
     history_hash
 }
 
-fn history_hash(history: String) -> HashMap<PathBuf, DirInfo> {
+pub fn history_hash(history: String) -> HashMap<PathBuf, DirInfo> {
     let mut hm: HashMap<PathBuf, DirInfo> =
         serde_json::from_str(&history).unwrap_or(HashMap::new());
     hm.retain(|k, _| k.exists());
     hm
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_history_hash_returns_empty_hashmap_if_dir_not_exist() {
-        let mut history: HashMap<PathBuf, DirInfo> = HashMap::new();
-        history.insert(PathBuf::from(r"\this\doesnt\exists"), DirInfo::new(1, true));
-        let result = history_hash(serde_json::to_string(&history).unwrap());
-        assert_eq!(result.is_empty(), true);
-    }
-
-    #[test]
-    fn test_history_hash_returns_empty_hashmap_if_dir_exists() {
-        let mut history: HashMap<PathBuf, DirInfo> = HashMap::new();
-        history.insert(dirs::home_dir().unwrap(), DirInfo::new(1, true));
-        let result = history_hash(serde_json::to_string(&history).unwrap());
-        assert_eq!(result.len(), 1);
-    }
 }
